Support composed withTranslation HOC.
This is quite pragmatic since compose is widely used in the React
community when dealing with HOCs. I don't think it is worth to support
very-specific implementations such as Lodash's `flow` and `flowRight`
arrays or to allow configuring custom names for the "compose" function.
Let's just see if this is good-enough for most use-cases.

Since compose interface doesn't seem to change that much across
libraries, this naive implementation should work with Redux, Recompose,
Lodash and probably many other implementations, as long as the user
doesn't rename the "compose" or "flow" or "flowRight" function to
something else.

Fixes #75