Fix bad extracted keys with Trans and basic HTML nodes.
This commit adds a better heuristic to detect which Trans component will
be converted to indices by react-i18next (when using
transKeepBasicHtmlNodes which is enabled by default). In particular, if
the node has many children or interpolation, react-i18next ignores the
transKeepBasicHtmlNodes option and converts the node to indices.

Fixes #108