Fix processing of spaces in Trans component.
Previously, if a JSXText node in a Trans component consisted in only one
whitespace, the extractor would strip it and ignore it for the indices
increment. This is wrong as react-i18next does take them into account.
However, react-i18next indeed ignores JSXText nodes that consist in only
whitespaces AND contains at list one linefeed. This commit makes the
extractor reflect the behavior of react-i18next.

Fixes #113