feat: add support for JSONv4
Also force using JSONv4 for ICU as it makes the plural categories more
correct. This is technically a breaking change, but as the feature was
explicitely marked as expermiental, I wouldn't worry too much.

I have yet to decide if JSONv4 should be made the default now (which
would require a major version increase). We surely want new users to use
JSONv4 by default, but a major version increase might be an opportunity
to include other changes.

closes #203