fix: don't use yarn to publish package
Recent versions of yarn refuse to publish packages that don't have a
yarn.lock and are located within an existing workspace. If we create an
empty yarn.lock, yarn still insists on locking the dependencies, which
we definitely don't want to include in the output package.

Always use npm to publish package. I guess using yarn for anything makes
little sense nowadays, so we might just want to drop it entirely at some
point.